Move expression utils from sql module to expr crate
